Update personal.js
There are several issues with this code. I will enumerate each one below.
Issue 1: You seem to be having issues with the idea of variable scope inside and outside of functions. The code as written would not run.
Issue 2: You either placed a function within an if statement or you added an else to the end of a function. In either case this will not work. It looks like you were working on adding a second option but didn't get it finished.